mqtt/requirements.txt
paho-mqtt
pymysql
omegaconf
geopy